Check user id in edit/remove, fix language

+++ app.py
@@ -1,5 +1,5 @@
 from flask import Flask
-from flask import redirect, render_template, request, session
+from flask import abort, redirect, render_template, request, session
 from werkzeug.security import generate_password_hash, check_password_hash
 import config
 import sqlite3
@@ -38,15 +38,19 @@ def create_entry():
 @app.route("/edit_entry/<int:entry_id>")
 def edit_entry(entry_id):
     entry = entries.get_entry(entry_id)
+    if entry["user_id"] != session["user_id"]:
+        abort(403)
     return render_template("edit_entry.html", entry=entry)
 
 @app.route("/update_entry", methods=["POST"])
 def update_entry():
+    entry_id = request.form["entry_id"]
+    entry = entries.get_entry(entry_id)
+    if entry["user_id"] != session["user_id"]:
+        abort(403)
     title = request.form["title"]
     artist = request.form["artist"]
     comment = request.form["comment"]
-    entry_id = request.form["entry_id"]
-
     entries.update_entry(title, artist, comment, entry_id)
 
     return redirect("/entry/" + str(entry_id))
@@ -55,10 +59,15 @@ def update_entry():
 def remove_entry(entry_id):
     if request.method == "GET":
         entry = entries.get_entry(entry_id)
+        if entry["user_id"] != session["user_id"]:
+            abort(403)
         return render_template("remove_entry.html", entry=entry)
 
     if request.method == "POST":
         if "remove" in request.form:
+            entry = entries.get_entry(entry_id)
+            if entry["user_id"] != session["user_id"]:
+                abort(403)
             entries.remove_entry(entry_id)
             return redirect("/")
         else:
